Add explanation for 'token_limit' parameter in README
An explanation about the 'token_limit' parameter in the `chunk` method of the `Chunker` class has been added to the README. This helps users understand how to control the size of each chunk of code. It also clarifies that the tokenization method and definition of a token depend on specific implementations and the type of content being processed.